znc: provide network name in password
We use the same ZNC server for many networks. In this case, the
password should be `login/network:password`. This is not something
that ZNC.el is handling. Therefore, we put the network name into the
login and we redefine `znc-erc-ssl-connector` to retrieve this network
name and build the correct password.

A cool side-effect is that the password is no longer displayed in the
messages.