Rework configuration to rely on MELPA and use-package
el-get had some advantages, like the setting of a minimal
configuration (no need of `:mode` most of the time). Also, quelpa
integration seems a bit fragile (it relies on an advice to be able to
use `:ensure t`). We'll see!

Some configuration has not been ported over. A bit of cleaning!